backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
sqlalchemy==2.0.43
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic[email]==2.11.7
python-dotenv==1.1.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# New dependencies for Phase 1
slowapi==0.1.9  # Rate limiting
python-json-logger==2.0.7  # JSON logging
prometheus-client==0.19.0  # Metrics